refactor: refactor isLinked to prevent exceptions on missing image or item
The refactoring of this method is to prevent exceptions being thrown
for a helper method to check if an item and image is linked. This
method is strictly for testing purposes and should not have any
unchecked exceptions. Hence, it is refactored in such a way that
even if there is no such item or image, it will return not linked.